rename piece_points.hpp to piece_position_points.hpp; piece_points lib still not renamed
