+++ src/core/piece_position_points/bpo_json_to_raw_json.cpp
@@ -1,4 +1,4 @@
-#include <piece_points.hpp>
+#include <piece_position_points.hpp>
 #include <piece_points_spec.hpp>
 
 using namespace piece_points;
